refactor: StaredRepositoryコンポーネントに分割した
